IdToken -> AccessToken in MSAL frontend code
Unable to authorize towards backend on localhost:5000 with IdToken, but it works with accessToken. With my limited understanding of idToken vs accessToken this makes sence as we are accessing resources from the backend API. However, I do not understand how this can work on other machines or in the dev environment.