Refactor syncing of selecting nodes between flow and mimir redux state.
Use .isSelected flag on nodes and edges to determine what to show in inspector.

Use react-flow internal state to determine what is currently selected in flow (for creating sub-project and alike).

Move FlowProvider to top of Mimir, allowing for accessing internal state throughout the application. Removes the need for FlowManipulator component.